Cleanup: Remove CMake, example code, libraries.json, presentations directory
The idea behind the example code and libraries.json file used by the
CMake code that this commit removes was originally to provide a
skeleton for a single distribution of Beman by using the CMake
FetchContent feature to acquire all the Beman libraries. This would be
done by keeping the libraries.json file up to date with every new
Beman library.

However, although there continues to be interest in a single Beman
distribution, the approach contained in this repository was never
followed up on and discussion on Discourse led to consensus for its
removal.

The code in the example/ directory has also been superseded by the
Beman examplar repository.

The presentations/ directory is also removed by this commit, since it
was not being kept up to date with the variety of presentations about
Beman that have been done since it was checked in.