Remove formatting rules from [LICENSE.APPROVED] and add [README.LICENSE] rule
After a discussion at the weekly sync meeting, we decided that the
header and footer required by [LICENSE.APPROVED] were superfluous and
potentially misleading, since it stated that "The Beman Project is
under the Apache License v2.0 with LLVM Exceptions" but projects were
not required to use that license.

To ensure users are properly informed about the project's license
after the removal of that header, this commit also adds a new rule
requiring the project's README file to specify the license used by the
project.